pyproject.toml
[tool.poetry]
name = "gvip"
version = "0.1.0"
description = ""
authors = ["John San Soucie"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
torch = "^1.12.0"
pyro-ppl = "^1.8.2"
mypy = "^0.971"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
coverage = "^6.4.4"
flake8 = "^5.0.4"
black = "^22.8.0"
isort = "^5.10.1"
pytest-cov = "^3.0.0"
codecov = "^2.1.12"
mypy = "^0.971"
gitchangelog = "^3.0.4"
mkdocs = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

